src/main/java/eRSPG/model/Reviewer.java
package eRSPG.model;

import java.time.LocalDateTime;

import javax.persistence.Column;
import javax.persistence.Entity;
import javax.persistence.Id;
import javax.persistence.Table;

@Entity
@Table(name="Reviewer")
public class Reviewer {
	@Id
	@Column(name="reviewerID")
	private int reviewerId;
	
	@Column(name="proposalID")
	private int proposalId;
	
	@Column(name="userID")
	private int userId;
	
	@Column(name="departmentID")
	private int departmentId;
	
	@Column(name="reviewComment")
	private String reviewComment;
	
	@Column(name="updatedAt")
	private LocalDateTime updatedAt;
	
	public Reviewer(){
		
	}

	public int getReviewerId() {
		return reviewerId;
	}

	public void setReviewerId(int reviewerId) {
		this.reviewerId = reviewerId;
	}

	public int getProposalId() {
		return proposalId;
	}

	public void setProposalId(int proposalId) {
		this.proposalId = proposalId;
	}

	public int getUserId() {
		return userId;
	}

	public void setUserId(int userId) {
		this.userId = userId;
	}

	public int getDepartmentId() {
		return departmentId;
	}

	public void setDepartmentId(int departmentId) {
		this.departmentId = departmentId;
	}

	public String getReviewComment() {
		return reviewComment;
	}

	public void setReviewComment(String reviewComment) {
		this.reviewComment = reviewComment;
	}

	public LocalDateTime getUpdatedAt() {
		return updatedAt;
	}

	public void setUpdatedAt(LocalDateTime updatedAt) {
		this.updatedAt = updatedAt;
	}
	
}
